+++ actionpack/CHANGELOG.md
@@ -86,4 +86,8 @@
 
     *Rafael Mendonça França*
 
+*   Implement a `:force` option for `cookies.delete` to allow deleting cookies not present in the request
+
+    *Felipe Zavan*
+
 Please check [7-1-stable](https://github.com/rails/rails/blob/7-1-stable/actionpack/CHANGELOG.md) for previous changes.